NewsSummaryTry2/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.0
openai==1.3.0
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0
